requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.6.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
contourpy==1.3.0
crayons==0.4.0
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deep-translator==1.11.4
distro==1.9.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==1.3.0
Flask==3.0.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
fugashi==1.3.2
fuzzywuzzy==0.18.0
gcsfs==2024.6.1
geographiclib==2.0
geopandas==1.0.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kaitaistruct==0.10
kiwisolver==1.4.7
libretranslatepy==2.1.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
narwhals==1.24.2
nest-asyncio==1.6.0
nordvpn-switcher==0.3.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.30.3
openpyxl==3.0.10
outcome==1.3.0
packaging==23.2
pandas==2.1.1
pandas-flavor==0.6.0
parse==1.19.1
parso==0.8.4
pathlib==1.0.1
patsy==0.5.6
pendulum==2.1.2
pillow==10.4.0
pingouin==0.5.5
plac==1.4.3
platformdirs==4.2.1
playwright==1.45.1
plotly==6.0.0
pminit==0.2.1
POI @ git+https://github.com/ptariindonesia/library_poi@1c273465a25877e654f0dc0ba344dc2e1d1b7893
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.24.4
psutil==5.9.6
pure-eval==0.2.2
pyarrow==19.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pydivert==2.1.0
pyee==11.1.0
Pygments==2.18.0
pyogrio==0.9.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyppeteer==1.0.2
pyproj==3.6.1
pyquery==2.0.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonmonkey==0.2.1
pytz==2023.3.post1
pytzdata==2020.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
random-user-agent==1.0.1
rapidfuzz==3.9.1
referencing==0.36.2
requests==2.31.0
requests-futures==1.0.1
requests-html==0.10.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
scraping @ git+https://github.com/ptariindonesia/Loketech_Cleaning.git@e679d02b27f75499dca93f397e2b1eaec652ea91
seaborn==0.13.2
selenium==4.27.1
sentry-sdk==2.3.1
setproctitle==1.3.3
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.41.1
style==1.1.0
tabulate==0.9.0
tenacity==9.0.0
thefuzz==0.22.1
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
unidic==1.1.0
update==0.0.1
urllib3==1.26.18
w3lib==2.1.2
wandb==0.17.0
wasabi==0.10.1
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.0
wsproto==1.2.0
xarray==2024.9.0
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
